+++ example_test.go
@@ -30,7 +30,7 @@ func Example() {
         //   "destination": {
         //     "endpoint": "https://otlp.io:443",
         //     "headers": {
-        //       "Access-Token": "aabbccdd...wwxxyyzz"
+        //       "access-token": "aabbccdd...wwxxyyzz"
         //     },
         //     "attributes": {
         //       "environment": "public",
@@ -59,7 +59,7 @@ func Example() {
         //   "diagnostics": {
         //     "endpoint": "https://otlp.io:443",
         //     "headers": {
-        //       "Access-Token": "wwxxyyzz...aabbccdd"
+        //       "access-token": "wwxxyyzz...aabbccdd"
         //     },
         //     "attributes": {
         //       "environment": "internal"
@@ -95,6 +95,7 @@ func Example() {
         //     "format": "json",
         //     "verbose": 1
         //   },
-	//   "disable_supervisor": false
+        //   "disable_supervisor": false,
+        //   "disable_diagnostics": false
         // }
 }